fix(migrate): move quick-migrate helpers out of the UI package
QuickMigrateJob, a background-job class in mihon.domain.migration,
imported normalizeQuickMigrateUrl/quickMigrateTargets/quickMigrateSkipped
from the UI layer (eu.kanade.tachiyomi.ui.browse.migration.manga), a
layering violation: the domain/background-job layer had a compile-time
dependency on a UI package. Move the three functions to
mihon.domain.migration.QuickMigrateHelpers.kt, which both the job and the
ViewModel now depend on.
