+++ app/src/main/java/eu/kanade/tachiyomi/ui/browse/migration/manga/MigrateMangaViewModel.kt
@@ -30,6 +30,7 @@ import logcat.LogPriority
 import mihon.core.common.utils.mutate
 import mihon.core.viewmodel.StateViewModel
 import mihon.domain.migration.QuickMigrateJob
+import mihon.domain.migration.quickMigrateTargets
 import tachiyomi.core.common.util.lang.launchIO
 import tachiyomi.core.common.util.system.logcat
 import tachiyomi.domain.manga.interactor.GetFavorites
@@ -267,29 +268,3 @@ sealed interface MigrationMangaEvent {
     data object FailedFetchingFavorites : MigrationMangaEvent
     data class QuickMigrateComplete(val count: Int, val removedCount: Int = 0) : MigrationMangaEvent
 }
-
-/** Leading-slash normalization matching how source urls are stored, for [targetSource]'s convention. */
-internal fun normalizeQuickMigrateUrl(url: String, targetSource: Source): String =
-    eu.kanade.tachiyomi.util.source.normalizeSourcePath(targetSource, url)
-
-/**
- * Pairs each selectable manga with its normalized target url, dropping the ones already favorited on
- * the target source. [existingFavoriteUrls] is the one-shot set of target-source favorite urls, so
- * duplicate detection is in-memory instead of one query per manga.
- */
-internal fun quickMigrateTargets(
-    selected: List<Manga>,
-    existingFavoriteUrls: Set<String>,
-    targetSource: Source,
-): List<Pair<Manga, String>> =
-    selected.mapNotNull { manga ->
-        val newUrl = normalizeQuickMigrateUrl(manga.url, targetSource)
-        if (newUrl in existingFavoriteUrls) null else manga to newUrl
-    }
-
-/** The other half of [quickMigrateTargets]: the entries the target source already has. */
-internal fun quickMigrateSkipped(
-    selected: List<Manga>,
-    existingFavoriteUrls: Set<String>,
-    targetSource: Source,
-): List<Manga> = selected.filter { normalizeQuickMigrateUrl(it.url, targetSource) in existingFavoriteUrls }
